pyproject.toml
[project]
name = "lemon-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
]


dependencies = [
  "aiocache>=0.12.3",
  "aioprometheus[starlette]>=23.12",
  "asyncpg==0.29",
  "bcrypt==4.0.1",
  "fastapi==0.117.1",
  "httpx>=0.28.1",
  "jinja2==3.1.2",
  "loguru==0.7",
  "passlib==1.7.4",
  "pydantic-settings>=2.10.1",
  "python-jose==3.3",
  "python-multipart==0.0.6",
  "python-ulid==1.1",
  "uvicorn==0.23.2",
  "validators==0.20",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.13.2",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.5",
    "testcontainers[postgresql]>=4.14.1",
]
